refactor: share the filter toolbar shell between accounts and transactions
