+++ frontend/src/pages/accounts/components/toolbar/FilterPanelBody.tsx
@@ -1,13 +1,13 @@
-import { useEffect, useId, useRef, useState } from 'react'
+import { useId, useState } from 'react'
 import { AnimatePresence, motion, useReducedMotion } from 'motion/react'
-import { Check, ChevronDown, X } from 'lucide-react'
+import { X } from 'lucide-react'
 import type { OptionItem } from '@/components/filters/OptionList'
 import { MultiSelectChecklist } from '@/components/filters/MultiSelectChecklist'
-import { getFilterOptionStyle } from '@/components/filters/optionAppearance'
+import { FacetSelectDropdown } from '@/components/list-controls/FacetSelectDropdown'
+import { FILTER_GLASS_SPRING } from '@/components/list-controls/toolbarStyles'
 import { joinClassNames } from '@/utils/classNames'
 import {
   FILTER_FACETS,
-  FILTER_GLASS_SPRING,
   type AccountFilterDraft,
   type FacetConfig,
   type FacetId,
@@ -51,7 +51,8 @@ export function FilterPanelBody({
   return (
     <div className={joinClassNames('contents', fillHeight && '!flex min-h-0 flex-1 flex-col')}>
       {mobile ? (
-        <MobileFacetSelect
+        <FacetSelectDropdown
+          facets={FILTER_FACETS}
           activeFacetId={activeFacetId}
           countFacet={draft.countFacet}
           onSelect={setActiveFacetId}
@@ -157,119 +158,6 @@ export function FilterPanelBody({
         </motion.div>
       )}
     </div>
-  )
-}
-
-type MobileFacetSelectProps = {
-  activeFacetId: FacetId
-  countFacet: (facet: FacetConfig) => number
-  onSelect: (facetId: FacetId) => void
-}
-
-/**
- * Renders the facet picker as a dropdown for the mobile full-screen panel, where the facet tab grid
- * is too cramped. The menu keeps the per-facet active-filter counts so the user can still tell which
- * facets carry filters without opening each one
- */
-function MobileFacetSelect({ activeFacetId, countFacet, onSelect }: MobileFacetSelectProps) {
-  const [open, setOpen] = useState(false)
-  const containerRef = useRef<HTMLDivElement>(null)
-  const activeFacet = FILTER_FACETS.find((facet) => facet.id === activeFacetId) ?? FILTER_FACETS[0]
-  const ActiveIcon = activeFacet.icon
-  const activeCount = countFacet(activeFacet)
-
-  // Close the menu on a pointer down outside it, so a tap on the checklist below dismisses the menu
-  // without also closing the whole modal
-  useEffect(() => {
-    if (!open) return undefined
-
-    function closeOnOutsidePointer(event: PointerEvent) {
-      if (containerRef.current && !containerRef.current.contains(event.target as Node)) setOpen(false)
-    }
-
-    document.addEventListener('pointerdown', closeOnOutsidePointer)
-    return () => document.removeEventListener('pointerdown', closeOnOutsidePointer)
-  }, [open])
-
-  return (
-    <div ref={containerRef} className="relative">
-      <button
-        type="button"
-        aria-haspopup="listbox"
-        aria-expanded={open}
-        className="app-input flex w-full items-center justify-between gap-2"
-        onClick={() => setOpen((isOpen) => !isOpen)}
-      >
-        <span className="flex min-w-0 items-center gap-2">
-          <ActiveIcon size={16} aria-hidden className="shrink-0" />
-          <span className="truncate">{activeFacet.label}</span>
-          {activeCount > 0 && <FacetCountBadge count={activeCount} />}
-        </span>
-        <ChevronDown
-          size={16}
-          aria-hidden
-          className="shrink-0 transition-transform duration-150"
-          style={{ transform: open ? 'rotate(180deg)' : 'none', color: 'var(--app-text-subtle)' }}
-        />
-      </button>
-
-      <AnimatePresence>
-        {open && (
-          <motion.ul
-            role="listbox"
-            className="absolute inset-x-0 top-full z-30 mt-1 max-h-[60vh] overflow-auto rounded-xl"
-            style={{
-              background: 'var(--app-input-bg)',
-              border: '1px solid var(--app-border-strong)',
-              boxShadow: 'var(--app-shadow-soft)',
-              backdropFilter: 'blur(16px)',
-            }}
-            initial={{ opacity: 0, y: -4 }}
-            animate={{ opacity: 1, y: 0 }}
-            exit={{ opacity: 0, y: -4 }}
-            transition={{ duration: 0.15 }}
-          >
-            {FILTER_FACETS.map((facet) => {
-              const FacetIcon = facet.icon
-              const facetCount = countFacet(facet)
-              const isActive = facet.id === activeFacetId
-              return (
-                <li
-                  key={facet.id}
-                  role="option"
-                  aria-selected={isActive}
-                  className="flex cursor-pointer items-center gap-2 px-4 py-2.5 text-sm transition-colors hover:bg-[var(--app-accent-soft)]"
-                  style={getFilterOptionStyle(isActive)}
-                  onClick={() => {
-                    onSelect(facet.id)
-                    setOpen(false)
-                  }}
-                >
-                  <FacetIcon size={16} aria-hidden className="shrink-0" />
-                  <span className="min-w-0 flex-1 truncate">{facet.label}</span>
-                  {facetCount > 0 && <FacetCountBadge count={facetCount} />}
-                  {isActive && <Check size={15} aria-hidden className="shrink-0" />}
-                </li>
-              )
-            })}
-          </motion.ul>
-        )}
-      </AnimatePresence>
-    </div>
-  )
-}
-
-/**
- * Renders the small accent pill showing how many filters a facet currently holds
- */
-function FacetCountBadge({ count }: { count: number }) {
-  return (
-    <span
-      className="flex h-[18px] min-w-[18px] shrink-0 items-center justify-center rounded-full px-1.5 text-[11px] font-medium"
-      style={{ background: 'var(--app-accent-soft)', color: 'var(--app-accent)' }}
-    >
-      {count}
-    </span>
   )
 }
