+++ frontend/src/pages/transactions/components/toolbar/FilterPanelBody.tsx
@@ -1,16 +1,16 @@
-import { useEffect, useId, useRef, useState } from 'react'
+import { useId, useState } from 'react'
 import { AnimatePresence, motion, useReducedMotion } from 'motion/react'
-import { Check, ChevronDown, X } from 'lucide-react'
+import { X } from 'lucide-react'
 import DateField from '@/components/date-field/DateField'
 import type { OptionItem } from '@/components/filters/OptionList'
 import { MultiSelectChecklist } from '@/components/filters/MultiSelectChecklist'
-import { getFilterOptionStyle } from '@/components/filters/optionAppearance'
+import { FacetSelectDropdown } from '@/components/list-controls/FacetSelectDropdown'
+import { FILTER_GLASS_SPRING } from '@/components/list-controls/toolbarStyles'
 import { joinClassNames } from '@/utils/classNames'
 import { formatMoneyInputLive, sanitizeMoneyInput } from '@/utils/moneyInput'
 import { ReferenceFacet } from '@/pages/transactions/components/toolbar/ReferenceFacet'
 import {
   FILTER_FACETS,
-  FILTER_GLASS_SPRING,
   type AmountDraft,
   type FacetConfig,
   type MultiSelections,
@@ -63,7 +63,8 @@ export function FilterPanelBody({
   return (
     <div className={joinClassNames('contents', fillHeight && '!flex min-h-0 flex-1 flex-col')}>
       {mobile ? (
-        <MobileFacetSelect
+        <FacetSelectDropdown
+          facets={FILTER_FACETS}
           activeFacetId={activeFacetId}
           countFacet={draft.countFacet}
           disabledFacetIds={disabledFacetIds}
@@ -190,123 +191,6 @@ export function FilterPanelBody({
         </motion.div>
       )}
     </div>
-  )
-}
-
-type MobileFacetSelectProps = {
-  activeFacetId: string
-  countFacet: (facet: FacetConfig) => number
-  disabledFacetIds: Set<string>
-  onSelect: (facetId: string) => void
-}
-
-/**
- * Renders the facet picker as a dropdown for the mobile full-screen panel, where the facet tab grid
- * is too cramped. The menu keeps the per-facet active-filter counts so the user can still tell which
- * facets carry filters without opening each one, and greys out any facet that is disabled
- */
-function MobileFacetSelect({ activeFacetId, countFacet, disabledFacetIds, onSelect }: MobileFacetSelectProps) {
-  const [open, setOpen] = useState(false)
-  const containerRef = useRef<HTMLDivElement>(null)
-  const activeFacet = FILTER_FACETS.find((facet) => facet.id === activeFacetId) ?? FILTER_FACETS[0]
-  const ActiveIcon = activeFacet.icon
-  const activeCount = countFacet(activeFacet)
-
-  // Close the menu on a pointer down outside it, so a tap on the editor below dismisses the menu
-  // without also closing the whole modal
-  useEffect(() => {
-    if (!open) return undefined
-
-    function closeOnOutsidePointer(event: PointerEvent) {
-      if (containerRef.current && !containerRef.current.contains(event.target as Node)) setOpen(false)
-    }
-
-    document.addEventListener('pointerdown', closeOnOutsidePointer)
-    return () => document.removeEventListener('pointerdown', closeOnOutsidePointer)
-  }, [open])
-
-  return (
-    <div ref={containerRef} className="relative">
-      <button
-        type="button"
-        aria-haspopup="listbox"
-        aria-expanded={open}
-        className="app-input flex w-full items-center justify-between gap-2"
-        onClick={() => setOpen((isOpen) => !isOpen)}
-      >
-        <span className="flex min-w-0 items-center gap-2">
-          <ActiveIcon size={16} aria-hidden className="shrink-0" />
-          <span className="truncate">{activeFacet.label}</span>
-          {activeCount > 0 && <FacetCountBadge count={activeCount} />}
-        </span>
-        <ChevronDown
-          size={16}
-          aria-hidden
-          className="shrink-0 transition-transform duration-150"
-          style={{ transform: open ? 'rotate(180deg)' : 'none', color: 'var(--app-text-subtle)' }}
-        />
-      </button>
-
-      <AnimatePresence>
-        {open && (
-          <motion.ul
-            role="listbox"
-            className="absolute inset-x-0 top-full z-30 mt-1 max-h-[60vh] overflow-auto rounded-xl"
-            style={{
-              background: 'var(--app-input-bg)',
-              border: '1px solid var(--app-border-strong)',
-              boxShadow: 'var(--app-shadow-soft)',
-              backdropFilter: 'blur(16px)',
-            }}
-            initial={{ opacity: 0, y: -4 }}
-            animate={{ opacity: 1, y: 0 }}
-            exit={{ opacity: 0, y: -4 }}
-            transition={{ duration: 0.15 }}
-          >
-            {FILTER_FACETS.map((facet) => {
-              const FacetIcon = facet.icon
-              const facetCount = countFacet(facet)
-              const isActive = facet.id === activeFacetId
-              const isDisabled = disabledFacetIds.has(facet.id)
-              return (
-                <li
-                  key={facet.id}
-                  role="option"
-                  aria-selected={isActive}
-                  aria-disabled={isDisabled}
-                  className={joinClassNames('flex items-center gap-2 px-4 py-2.5 text-sm transition-colors', isDisabled ? 'cursor-not-allowed' : 'cursor-pointer hover:bg-[var(--app-accent-soft)]')}
-                  style={{ ...getFilterOptionStyle(isActive), opacity: isDisabled ? 0.4 : 1 }}
-                  onClick={() => {
-                    if (isDisabled) return
-                    onSelect(facet.id)
-                    setOpen(false)
-                  }}
-                >
-                  <FacetIcon size={16} aria-hidden className="shrink-0" />
-                  <span className="min-w-0 flex-1 truncate">{facet.label}</span>
-                  {!isDisabled && facetCount > 0 && <FacetCountBadge count={facetCount} />}
-                  {isActive && <Check size={15} aria-hidden className="shrink-0" />}
-                </li>
-              )
-            })}
-          </motion.ul>
-        )}
-      </AnimatePresence>
-    </div>
-  )
-}
-
-/**
- * Renders the small accent pill showing how many filters a facet currently holds
- */
-function FacetCountBadge({ count }: { count: number }) {
-  return (
-    <span
-      className="flex h-[18px] min-w-[18px] shrink-0 items-center justify-center rounded-full px-1.5 text-[11px] font-medium"
-      style={{ background: 'var(--app-accent-soft)', color: 'var(--app-accent)' }}
-    >
-      {count}
-    </span>
   )
 }
